Take out QuestionLayout
