+++ src/components/QuestionTextInput.jsx
@@ -1,15 +1,10 @@
-import { useState, useEffect, useContext } from "react";
+import { useState, useContext } from "react";
 
 import styles from "./QuestionTextInput.module.css";
 
-import Card from "lib/card/Card";
-import QuestionBackCard from "./QuestionBackCard";
-import QuestionCardLayout from "lib/card/QuestionCardLayout";
-import ContentGridLayout from "lib/card/ContentGridLayout";
-import FancyHeader from "lib/FancyHeader";
 import Button from "lib/Button";
 import { CardModalContext } from "lib/card/CardModal";
-import CardHint from "./CardHint";
+import QuestionLayout from "./QuestionLayout";
 
 function QuestionTextInput({
   title,
@@ -25,12 +20,6 @@ function QuestionTextInput({
   const [answerState, setAnswerState] = useState(alreadySolved);
   const [answerValue, setAnswerValue] = useState("");
 
-  useEffect(() => {
-    if (alreadySolved) {
-      ctx.flip(false);
-    }
-  }, []);
-
   const ctx = useContext(CardModalContext);
 
   const checkAnswer = () => {
@@ -43,49 +32,30 @@ function QuestionTextInput({
     ctx.flip();
   };
 
-  const content = (
-    <div className={styles["content-wrapper"]}>
-      {children}
-      <div className={styles["form"]}>
-        <input
-          className={styles["input"]}
-          placeholder="Введите ответ"
-          type="text"
-          value={answerValue}
-          onChange={(e) => setAnswerValue(e.target.value)}
-        />
-        <Button onClick={checkAnswer}>Отправить</Button>
-      </div>
-    </div>
-  );
-
-  const front = (
-    <QuestionCardLayout>
-      <ContentGridLayout
-        title={
-          <div className={styles["heading"]}>
-            <FancyHeader className={styles["title"]} text={title} />
-            <CardHint hint={hint} />
-          </div>
-        }
-        content={content}
-      />
-    </QuestionCardLayout>
-  );
-
   return (
-    <Card
-      front={front}
-      back={
-        <QuestionBackCard
-          info={info}
-          showSecondRowImages={showSecondRowImages}
-          images={images}
-          answerState={answerState}
-          title={title}
-        />
-      }
-    />
+    <QuestionLayout
+      title={title}
+      id={id}
+      answerState={answerState}
+      images={images}
+      hint={hint}
+      info={info}
+      showSecondRowImages={showSecondRowImages}
+    >
+      <div className={styles["content-wrapper"]}>
+        {children}
+        <div className={styles["form"]}>
+          <input
+            className={styles["input"]}
+            placeholder="Введите ответ"
+            type="text"
+            value={answerValue}
+            onChange={(e) => setAnswerValue(e.target.value)}
+          />
+          <Button onClick={checkAnswer}>Отправить</Button>
+        </div>
+      </div>
+    </QuestionLayout>
   );
 }
