+++ src/components/QuestionTextInput.module.css
@@ -4,20 +4,6 @@
   justify-content: space-evenly;
 }
 
-.heading {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.title {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  font-size: 3.25vmax;
-  composes: noselect from "lib/commons.module.css";
-}
-
 .form {
   margin-top: 3rem;
   display: flex;
